enemy spawner progress
Made a Marker2D Spawner node that can be placed on any scene to spawn enemies. It spawns a new enemy when you press the spacebar for now.

The spawner needs to preload the enemies with their file path for it work.

It has 2 exported values that can be used to customize it.
Spawner Type lets you select which color to spawn from random, red, green or blue. Spawn Radius Size determines how far the enemies can spawn from the spawner.